requirements.txt
azure-core==1.24.1
azure-cosmos==4.3.0
azure-functions==1.11.2
botocore==1.27.10
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
s3transfer==0.6.0
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
